tumor_classifier/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
tensorflow-cpu==2.13.0
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
scikit-learn==1.3.2
requests==2.31.0 